+++ development_tools/compose.yml
@@ -58,4 +58,3 @@ services:
       interval: 10s
       timeout: 10s
       retries: 3
-
