+++ src/hct_mis_api/apps/sanction_list/strategies/__init__.py
@@ -1,0 +1,3 @@
+from .registry import registry
+
+__all__ = ["registry"]
