+++ src/hct_mis_api/apps/sanction_list/strategies/_base.py
@@ -1,0 +1,12 @@
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from ..models import SanctionList
+
+
+class BaseSanctionList:
+    def __init__(self, context: "SanctionList") -> None:
+        self.context = context
+
+    def refresh(self) -> None:
+        raise NotImplementedError()
