+++ src/hct_mis_api/apps/sanction_list/strategies/eu.py
@@ -1,0 +1,5 @@
+from ._base import BaseSanctionList
+
+
+class EUSanctionList(BaseSanctionList):
+    pass
