+++ src/hct_mis_api/apps/sanction_list/strategies/registry.py
@@ -1,0 +1,10 @@
+from strategy_field.registry import Registry
+
+from ._base import BaseSanctionList
+
+
+class SanctionListRegistry(Registry):
+    pass
+
+
+registry = SanctionListRegistry(BaseSanctionList)
