+++ tests/unit/conftest.py
@@ -2,6 +2,7 @@ import logging
 import os
 import re
 import sys
+from pathlib import Path
 from time import sleep
 from typing import Any
 
@@ -35,6 +36,9 @@ def clear_default_cache() -> None:
 
 def pytest_configure(config: Config) -> None:
     pytest.localhost = True if config.getoption("--localhost") else False
+    here = Path(__file__).parent
+    utils = here.parent / "extras"
+    sys.path.append(str(utils))
 
     sys._called_from_pytest = True
     from django.conf import settings
