+++ tox.ini
@@ -42,8 +42,8 @@ commands =
         --reruns-delay 1 \
         --randomly-seed=42 \
         --create-db \
-        {posargs:} \
-        ./tests/unit
+        {posargs: tests/unit}
+
 
 [testenv:selenium]
 runner = uv-venv-lock-runner
